Added dir read/update/delete checks

+++ HomeCloud-Server/Controllers/DirectoryController.cs
@@ -23,6 +23,13 @@ namespace HomeCloud_Server.Controllers
             _configService = configurationService;
         }
 
+        private User GetUser()
+        {
+            string token = Request.Headers["ApiKey"];
+            User user = _databaseService.GetUserFromToken(token);
+            return user;
+        }
+
         /// <summary>
         /// Creates a directory, specifying a name and parent
         /// </summary>
@@ -32,8 +39,7 @@ namespace HomeCloud_Server.Controllers
         [HttpGet("CreateDirectory")]
         public async Task<IActionResult> CreateDirectory(string DirectoryName, uint ParentDirectoryID = 0)
         {
-            string token = Request.Headers["ApiKey"];
-            User user = _databaseService.GetUserFromToken(token);
+            User user = GetUser();
             if (PermissionChecker.AllowedToCreate(ParentDirectoryID, user.UserID, _databaseService))
             {
                 //Create folder object
@@ -41,7 +47,8 @@ namespace HomeCloud_Server.Controllers
                 {
                     ParentDirectory = ParentDirectoryID,
                     DirectoryName = DirectoryName,
-                    CreatedOn = (ulong)DateTime.UtcNow.Subtract(DateTime.UnixEpoch).TotalSeconds //Current Time
+                    CreatedOn = (ulong)DateTime.UtcNow.Subtract(DateTime.UnixEpoch).TotalSeconds, //Current Time
+                    OwnerID = user.UserID
                 };
                 //Add folder to the database
                 await _databaseService.CreateNewDirectory(d);
@@ -64,7 +71,7 @@ namespace HomeCloud_Server.Controllers
         public async Task<IActionResult> GetAllDirectories()
         {
             List<Models.Directory> d = await _databaseService.GetDirectoriesAsync();
-            return Ok(d);
+            return Ok(d); 
         }
 
         /// <summary>
@@ -75,22 +82,40 @@ namespace HomeCloud_Server.Controllers
         [HttpGet("GetSubdirectories")]
         public async Task<IActionResult> GetSubdirectories(uint ParentDirectoryID)
         {
-            List<Models.Directory> d = await _databaseService.GetSubdirectoriesAsync(ParentDirectoryID);
-            return Ok(d);
+            User user = GetUser();
+            if (PermissionChecker.AllowedToView(user, ParentDirectoryID, _databaseService))
+            {
+                List<Models.Directory> d = await _databaseService.GetSubdirectoriesAsync(ParentDirectoryID);
+                return Ok(d); 
+            }
+            else
+            {
+                return Unauthorized();
+            }
         }
 
         [HttpGet("GetFilesInDirectory")]
         public async Task<IActionResult> GetFilesInDirectory(uint DirectoryID)
         {
-            List<Models.File> f = await _databaseService.GetAllFilesInDirAsync(DirectoryID);
-            return Ok(f);
+            User user = GetUser();
+            if (PermissionChecker.AllowedToView(user, DirectoryID, _databaseService))
+            {
+                List<Models.File> f = await _databaseService.GetAllFilesInDirAsync(DirectoryID);
+                return Ok(f); 
+            }
+            else { return Unauthorized(); }
         }
 
         [HttpGet("GetContents")]
         public async Task<IActionResult> GetContents(uint DirectoryID)
         {
-            Models.DirectoryContents dc = await _databaseService.GetDirectoryContents(DirectoryID);
-            return Ok(dc);
+            User user = GetUser();
+            if (PermissionChecker.AllowedToView(user, DirectoryID, _databaseService))
+            {
+                Models.DirectoryContents dc = await _databaseService.GetDirectoryContents(DirectoryID);
+                return Ok(dc); 
+            }
+            else { return Unauthorized(); }
         }
 
         /// <summary>
@@ -101,8 +126,13 @@ namespace HomeCloud_Server.Controllers
         [HttpGet("RenameDirectory")]
         public async Task<IActionResult> RenameDirectory(uint DirectoryID, string NewName)
         {
-            await _databaseService.RenameDirectoryAsync(DirectoryID, NewName);
-            return Ok();
+            User user = GetUser();
+            if (PermissionChecker.AllowedToEdit(DirectoryID, user.UserID, _databaseService))
+            {
+                await _databaseService.RenameDirectoryAsync(DirectoryID, NewName);
+                return Ok(); 
+            }
+            else { return Unauthorized(); }
         }
 
 
@@ -114,23 +144,28 @@ namespace HomeCloud_Server.Controllers
         [HttpDelete("DeleteDirectory")]
         public async Task<IActionResult> DeleteDirectory(uint DirectoryID)
         {
-            //Check dir is empty
-            //no files contained
-            List<Models.Directory> containedDirs = await _databaseService.GetSubdirectoriesAsync(DirectoryID);
-            //no directories contained
-            List<Models.File> containedFiles = await _databaseService.GetAllFilesInDirAsync(DirectoryID);
+            User user = GetUser();
+            if (PermissionChecker.AllowedToDelete(DirectoryID, user.UserID, _databaseService))
+            {
+                //Check dir is empty
+                //no files contained
+                List<Models.Directory> containedDirs = await _databaseService.GetSubdirectoriesAsync(DirectoryID);
+                //no directories contained
+                List<Models.File> containedFiles = await _databaseService.GetAllFilesInDirAsync(DirectoryID);
 
-            //Check empty
-            if(containedDirs.Count == 0 && containedFiles.Count == 0)
-            {
-                //Delete it
-                await _databaseService.DeleteDirectoryAsync(DirectoryID);
-                return Ok();
+                //Check empty
+                if (containedDirs.Count == 0 && containedFiles.Count == 0)
+                {
+                    //Delete it
+                    await _databaseService.DeleteDirectoryAsync(DirectoryID);
+                    return Ok();
+                }
+                else
+                {
+                    return Conflict("There were files or folders present in the directory. Please remove these and try again");
+                } 
             }
-            else
-            {
-                return Conflict("There were files or folders present in the directory. Please remove these and try again");
-            }
+            else { return Unauthorized(); }
         }
     }
 }
